rework scraperapi folder
